+++ react/__mocks__/vtex.styleguide.js
@@ -1,7 +1,11 @@
-import React, { Fragment } from 'react'
+import React from 'react'
 
 export function Button(props) {
   return <button type="button"> {props.children} </button>
+}
+
+export function Input(props) {
+  return <input {...props} />
 }
 
 export function withToast(Comp) {