+++ CHANGELOG.md
@@ -9,8 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- Snapshot tests for ProductPrice.
-- Snapshot tests for `BuyButton`
+- Snapshot tests for `ProductPrice`.
+- Snapshot tests for `BuyButton`.
+- Snapshot tests for `AvailabilitySubscriber`.
 
 ## [3.17.2] - 2019-02-18